First minimal changes, which fix compiler errors
These are changes to the original C# which was auto-converted from VB.

These are the minimum set of changes which were needed to build (and
run) successfully, including:

- Use generic List<> instead of ArrayList
- Implicit char to string conversions made explicit
- Implicit Token.Data object to string conversions made explicit
- Implicit enum to int conversions made explicit
- Use "()" when invoking a method
- Various int to short and short to ushort conversions
- "using Microsoft.VisualBasic" to resolve Strings.ascW and Strings.ChrW

This commit does NOT include any of the following types of change:

- Refactoring
- New features
- Changes to the oublic API

The only public API difference so far, is that this project implements
"namespace GOLDEngine" instead of "namespace GOLD".

I changed the namespace name, because I expect to make some further
breaking changes to the public API, in future commits.